feat(header): moved child components to story. context added to manage state across child components
