+++ src/components/Navigation/Navigation.tsx
@@ -6,14 +6,6 @@ import { HeaderContext } from '../Header/Header';
 export interface NavigationProps extends React.HTMLAttributes<HTMLElement> {
   backBtnLabel?: string;
   /**
-   * Optional label
-   */
-  label?: string;
-  /**
-   * Optional onBack function
-   */
-  onBack?: () => void;
-  /**
    * Optional visible state
    */
   visible?: boolean;
@@ -24,11 +16,10 @@ const Navigation = ({
   children,
   className,
   id,
-  label = 'Main Menu',
-  onBack,
   visible = true,
 }: React.PropsWithChildren<NavigationProps>) => {
-  const { expanded } = React.useContext(HeaderContext);
+  const { expanded, expandedItem, setExpandedItem } = React.useContext(HeaderContext);
+  const onBack = () => setExpandedItem(expandedItem === 'Main Menu' ? expandedItem : 'Main Menu');
   return (
     <nav
       aria-labelledby={`${id}-label`}
@@ -44,7 +35,7 @@ const Navigation = ({
         className={`${px}-nav__label`}
         style={{ '--visible': visible ? 'visible' : 'hidden' } as React.CSSProperties}
       >
-        {label}
+        {expandedItem}
       </h2>
       {children}
     </nav>
