--- src/components/Navigation/NavigationItem/NavigationItem.test.ts
@@ -1,13 +1,0 @@
-// import { render, screen } from '@testing-library/react';
-// import NavigationItem from './NavigationItem';
-
-// describe('NavigationItem', () => {
-//     it('renders the navigation item correctly', () => {
-//         render(<NavigationItem href="/" label="Home" />);
-        
-//         const navigationItem = screen.getByText('Home');
-//         expect(navigationItem).toBeInTheDocument();
-//     });
-
-    
-// });
