+++ src/components/Search/_search.scss
@@ -10,9 +10,9 @@
 
   &__overlay {
     background-color: rgba(0, 0, 0, 50%);
-    inset: 0;
     display: none;
     height: 100%;
+    inset: 0;
     position: fixed;
     width: 100%;
     z-index: 2;
